Added ability to mark the band gap from a calculation. This is done for both spin channels.
Added ability to shift the eigenvalues and mark the gap on the bandstructures.

Shifting of eigenvalues can now be done other than just the Fermi energy (through zero_vbm, zero_cbm and zero_shift) options
when initialising the class.

mark_gap options to mark the band gap in the band structure. Can be done for each spin channel.

Added ability to set kpoint limit by high symmetry labels or by kpoint index.

Tidied up the if statement for the y-axis label when plotting a bandstructure.

Added the necessary labels if we offset such that the zero of energy lies along the VBM/CBM.