Corrected check for high-symmetry points.
We now take the mod instead of absolute value (as fractional coordinates) can
wrap around the cell boundary.

This should fix test cases such as for the H-point of BCC lattice which is stored as
H = 0.5 -0.5 0.5
and if we have a k-point
k = 0.5 0.5 0.5,
subtracting the two and taking the absolute value will give
|H - k| = 0 1 0
and the if statement wrongly does not recognise it as a high-symmetry point.